Apply feedbacks
Changes:
 - Removes 'example_' prefix from examples.
 - Refactors example_proxy.
 - Removes tizen_log dependency from example_stub.
