--- packages/tizen_rpc_port/example/example_proxy/tizen/.gitignore
@@ -1,5 +1,0 @@
-flutter/
-.vs/
-*.user
-bin/
-obj/
