+++ packages/tizen_rpc_port/example/proxy/pubspec.yaml
@@ -1,5 +1,5 @@
-name: example_stub
-description: Demonstrates how to use the rpc_port_stub plugin.
+name: example_proxy
+description: Demonstrates how to use the tizen_rpc_port proxy plugin.
 
 publish_to: none
