+++ packages/tizen_rpc_port/example/proxy/tizen/shared/res/ic_launcher.png
