+++ packages/tizen_rpc_port/example/proxy/tizen/tizen-manifest.xml
