+++ packages/tizen_rpc_port/example/stub/tizen/shared/res/ic_launcher.png
