+++ packages/tizen_rpc_port/example/stub/tizen/tizen-manifest.xml
